+++ drivers/wifi/siwx91x/CMakeLists.txt
@@ -5,6 +5,7 @@ if(CONFIG_WIFI_SILABS_SIWX91X)
   zephyr_library_sources(
     siwx91x_wifi.c
     siwx91x_wifi_ap.c
+    siwx91x_wifi_sta.c
     siwx91x_wifi_scan.c
     siwx91x_wifi_ps.c)
